track local lz4 project
Summary:
This will be a very important compression algorithm for future
development.

Here are some quick and rough benches I did with png image decoding
against zlib (used throughout the AOSP currently, and miniz is used
throughout cfXe).

The initial testing I've done with lz4 over zlib was with png images.
The "default compression" was used for all algorithms. I manually
decompressed and re-compressed all non-zlib types for quick testing,
and mean averaged 100 runs of each for decompressing on a test build.
The ratios are not as good as zlib for any of these, but the speed gain
(and lower load overhead) on a class 20 emmc (ext4 android defaults)
is quite good.

   zlib (default libpng): 247ms
   snappy: 54ms
   lz4: 38ms

Test Plan: verified

Reviewers: randall.rushing

Differential Revision: http://review.codefirex.com/D19